Fixed isLeaf predicate
"You are protected, in short, by your ability to love! The only
protection that can possibly work against the lure of power like
Voldemort's! In spite of all the temptation you have endured, all the
suffering, you remain pure of heart, just as pure as you were at the
age of eleven, when you stared into a mirror that reflected your
heart's desire, and it showed you only the way to thwart Lord
Voldemort, and not immortality or riches."
—Albus Dumbledore to Harry. [Harry Potter]